Irfan - Fixed quantization bug. Redirected entropy output to testdata directory.

+++ Encoder/Encoder.cpp
@@ -184,7 +184,7 @@ int main(int argCnt, char **args)
 	 int8_t** DEC_RES_FRAME_2D	= new  int8_t*[height];
 	int32_t** DEC_TC_FRAME_2D	= new int32_t*[height];
 	int32_t** QTC_FRAME_2D		= new int32_t*[height];
-	uint8_t** QP_FRAME_2D		= new uint8_t*[height];
+	uint32_t** QP_FRAME_2D		= new uint32_t*[height];
 
 	for (unsigned int row = 0; row < height; row++) {
 		CUR_FRAME_2D[row] = new uint8_t[width];
@@ -204,7 +204,7 @@ int main(int argCnt, char **args)
 		DEC_RES_FRAME_2D[row] = new  int8_t[width];
 		DEC_TC_FRAME_2D[row] = new int32_t[width];
 		QTC_FRAME_2D[row] = new int32_t[width];
-		QP_FRAME_2D[row] = new uint8_t[width];
+		QP_FRAME_2D[row] = new uint32_t[width];
 	}
 
 	
@@ -242,7 +242,7 @@ int main(int argCnt, char **args)
 	int8_t** DEC_RES_FRAME_2DS = new  int8_t*[height];
 	int32_t** DEC_TC_FRAME_2DS = new int32_t*[height];
 	int32_t** QTC_FRAME_2DS = new int32_t*[height];
-	uint8_t** QP_FRAME_2DS = new uint8_t*[height];
+	uint32_t** QP_FRAME_2DS = new uint32_t*[height];
 
 	for (unsigned int row = 0; row < height; row++) {
 		CUR_FRAME_2DS[row] = new uint8_t[width];
@@ -262,7 +262,7 @@ int main(int argCnt, char **args)
 		DEC_RES_FRAME_2DS[row] = new  int8_t[width];
 		DEC_TC_FRAME_2DS[row] = new int32_t[width];
 		QTC_FRAME_2DS[row] = new int32_t[width];
-		QP_FRAME_2DS[row] = new uint8_t[width];
+		QP_FRAME_2DS[row] = new uint32_t[width];
 	}
